Components: Extract the region navigation into a reusable Higher-Order Component
