+++ editor/layout/index.js
@@ -7,8 +7,7 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
-import { Component } from '@wordpress/element';
-import { NoticeList, Popover, KeyboardShortcuts } from '@wordpress/components';
+import { NoticeList, Popover, navigateRegions } from '@wordpress/components';
 
 /**
  * Internal dependencies
@@ -29,77 +28,29 @@ import {
 	getNotices,
 } from '../selectors';
 
-class Layout extends Component {
-	constructor() {
-		super( ...arguments );
-		this.bindContainer = this.bindContainer.bind( this );
-		this.focusNextRegion = this.focusRegion.bind( this, 1 );
-		this.focusPreviousRegion = this.focusRegion.bind( this, -1 );
-		this.onClick = this.onClick.bind( this );
-		this.state = {
-			isFocusingRegions: false,
-		};
-	}
+function Layout( { mode, isSidebarOpened, notices, ...props } ) {
+	const className = classnames( 'editor-layout', {
+		'is-sidebar-opened': isSidebarOpened,
+	} );
 
-	bindContainer( ref ) {
-		this.container = ref;
-	}
-
-	focusRegion( offset ) {
-		const regions = [ ...this.container.querySelectorAll( '[role="region"]' ) ];
-		if ( ! regions.length ) {
-			return;
-		}
-		let nextRegion = regions[ 0 ];
-		const selectedIndex = regions.indexOf( document.activeElement );
-		if ( selectedIndex !== -1 ) {
-			let nextIndex = selectedIndex + offset;
-			nextIndex = nextIndex === -1 ? regions.length - 1 : nextIndex;
-			nextIndex = nextIndex === regions.length ? 0 : nextIndex;
-			nextRegion = regions[ nextIndex ];
-		}
-
-		nextRegion.focus();
-		this.setState( { isFocusingRegions: true } );
-	}
-
-	onClick() {
-		this.setState( { isFocusingRegions: false } );
-	}
-
-	render() {
-		const { mode, isSidebarOpened, notices, ...props } = this.props;
-		const className = classnames( 'editor-layout', {
-			'is-sidebar-opened': isSidebarOpened,
-			'is-focusing-regions': this.state.isFocusingRegions,
-		} );
-
-		// Disable reason: Clicking the editor should dismiss the regions focus style
-		/* eslint-disable jsx-a11y/no-static-element-interactions, jsx-a11y/onclick-has-role, jsx-a11y/click-events-have-key-events */
-		return (
-			<div key="editor" className={ className } ref={ this.bindContainer } onClick={ this.onClick }>
-				<DocumentTitle />
-				<NoticeList onRemove={ props.removeNotice } notices={ notices } />
-				<UnsavedChangesWarning />
-				<AutosaveMonitor />
-				<Header />
-				<div className="editor-layout__content">
-					<div className="editor-layout__editor">
-						{ mode === 'text' && <TextEditor /> }
-						{ mode === 'visual' && <VisualEditor /> }
-					</div>
-					<MetaBoxes location="normal" />
+	return (
+		<div className={ className }>
+			<DocumentTitle />
+			<NoticeList onRemove={ props.removeNotice } notices={ notices } />
+			<UnsavedChangesWarning />
+			<AutosaveMonitor />
+			<Header />
+			<div className="editor-layout__content">
+				<div className="editor-layout__editor">
+					{ mode === 'text' && <TextEditor /> }
+					{ mode === 'visual' && <VisualEditor /> }
 				</div>
-				{ isSidebarOpened && <Sidebar /> }
-				<Popover.Slot />
-				<KeyboardShortcuts shortcuts={ {
-					'ctrl+r': this.focusNextRegion,
-					'ctrl+shift+r': this.focusPreviousRegion,
-				} } />
+				<MetaBoxes location="normal" />
 			</div>
-			/* eslint-enable jsx-a11y/no-static-element-interactions, jsx-a11y/onclick-has-role, jsx-a11y/click-events-have-key-events */
-		);
-	}
+			{ isSidebarOpened && <Sidebar /> }
+			<Popover.Slot />
+		</div>
+	);
 }
 
 export default connect(
@@ -109,4 +60,4 @@ export default connect(
 		notices: getNotices( state ),
 	} ),
 	{ removeNotice }
-)( Layout );
+)( navigateRegions( Layout ) );
