+++ editor/layout/style.scss
@@ -5,17 +5,6 @@
 
 .editor-layout {
 	position: relative;
-
-	&.is-focusing-regions [role="region"]:focus:after {
-		content: '';
-		position: absolute;
-		top: 0;
-		bottom: 0;
-		left: 0;
-		right: 0;
-		pointer-events: none;
-		border: 2px solid $blue-medium-400;
-	}
 }
 
 .editor-layout__content {
